fix : merge 진행

+++ src/main/java/com/kosa/moimeasy/user/service/RoleService.java
@@ -1,7 +1,7 @@
 package com.kosa.moimeasy.user.service;
 
+import com.kosa.moimeasy.security.repository.RoleRepository;
 import com.kosa.moimeasy.user.entity.Role;
-import com.kosa.moimeasy.user.repository.RoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.Optional;
@@ -12,7 +12,7 @@ public class RoleService {
     @Autowired
     private RoleRepository roleRepository;
 
-    public Optional<Role> findById(Integer roleId) {
+    public Optional<Role> findById(long roleId) {
         return roleRepository.findById(roleId);
     }
 }
